// types.ts
export interface Book {
    id: string;
    author: string;
    audioLink: string;
    averageRating: number;
    bookDescription: string;
    authorDescription: string;
    imageLink: string;
    keyIdeas: number;
    status: string;
    subTitle: string;
    subscriptionRequired: boolean;
    summary: string;
    tags: string[];
    title: string;
    totalRating: number;
    type: string;

  }